Move checkPayloadIs out of the public API
